продолжаю работы с фрагментом коллекций 13 осталась переработка таблицы movie - хранить больше данных в том числе путь к сохраненным фото на диске

+++ app/src/main/java/org/sniffsnirr/skillcinema/usecases/DeleteMovieFromCollectionUsecase.kt
@@ -8,20 +8,31 @@ import java.io.File
 import javax.inject.Inject
 
 @ActivityRetainedScoped
-class DeleteMovieFromCollectionUsecase @Inject constructor(val databaseRepository: DatabaseRepository,
-                                                           val getOneMovieFromDBByCollectionUsecase:GetOneMovieFromDBByCollectionUsecase) {
+class DeleteMovieFromCollectionUsecase @Inject constructor(
+    val databaseRepository: DatabaseRepository,
+    val getOneMovieFromDBByCollectionUsecase: GetOneMovieFromDBByCollectionUsecase,
+    val getCountMovieInCollection:GetCountMovieInCollection
+) {
 
-    suspend fun deleteMovieFromCollection(movieRVModel: MovieRVModel, collectionId: Long,dir: File){
-
-        //удаляю файл
-        val movieForDeleteFromCollection=getOneMovieFromDBByCollectionUsecase.getMovie(movieRVModel.kinopoiskId!!.toLong(),collectionId)
-        try {
-            val file=File(dir, movieForDeleteFromCollection?.poster!!)
-            file.delete()
-        } catch (e: Exception) {
-            Log.d("DeleteFileError","${e.printStackTrace()}")
+    suspend fun deleteMovieFromCollection(movieRVModel: MovieRVModel, collectionId: Long) {
+        if(getCountMovieInCollection.isAlreadyExist(movieRVModel.kinopoiskId!!.toLong(),collectionId)) {
+            //удаляю файл
+            val movieForDeleteFromCollection = getOneMovieFromDBByCollectionUsecase.getMovie(
+                movieRVModel.kinopoiskId!!.toLong(),
+                collectionId
+            )
+            try {
+                //  val file=File(dir, movieForDeleteFromCollection?.poster!!)
+                val file = File(movieForDeleteFromCollection?.poster!!)
+                file.delete()
+            } catch (e: Exception) {
+                Log.d("DeleteFileError", "${e.printStackTrace()}")
+            }
+            //удаляю из БД
+            databaseRepository.deleteMovieFromCollection(
+                movieRVModel.kinopoiskId!!.toLong(),
+                collectionId
+            )
         }
-        //удаляю из БД
-        databaseRepository.deleteMovieFromCollection(movieRVModel.kinopoiskId!!.toLong(),collectionId)
     }
 }
